requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
openai==1.3.0
python-multipart==0.0.6
pydantic==2.5.2
huggingface-hub==0.16.4 